+++ docs/examples/getting_started/named_tuple_controller.py
@@ -16,6 +16,7 @@ class UserModel(NamedTuple):
 
 class UserController(
     Controller[PydanticSerializer],
+    # Dataclasses do not support field name aliases, so can't use Headers :(
 ):
     def post(self, parsed_body: Body[UserCreateModel]) -> UserModel:
         return UserModel(uid=uuid.uuid4(), email=parsed_body.email)