+++ CHANGELOG.md
@@ -22,6 +22,10 @@ No breaking changes will be made.
 - Fixed `StreamingValidator` swallowing errors
   when `validate_events` was `True`, but no event model was resolved
 
+### Misc
+
+- Added `QuerySet` tutorial
+
 
 ## Version 0.4.0 (2026-03-29)